fix(population): Math.Round statt Floor/Ceiling + Stub-Seam für BuildPlanForTick
Root Cause: IEEE-754-Float-Präzision (0.7f = 0.699999988…) verfälschte
Floor(10*0.7f)=6 statt 7 und Ceiling(10*1.15f)=12 statt 11.

- PopulationLedger.ApplyDailyGrowthTick: Ceiling → Math.Round(double)
- PopulationLedger.GetRevengeQuota: Floor → Math.Round(double)
- NoGameStoryDirector.RecomputeRevengeAfterDayTickStub: Floor → Math.Round(double)
- InfectedRaidSpawnService.BuildPlanForTick: StubDirector-Seam bei
  Current.Game==null (Regression-Tests ohne Live-Game)
- HordeManifest T20: HitCount-Erwartung auf Math.Round-Verhalten angepasst

Ergebnis: 8/8 vorbestehende Test-Failures behoben.
- TutorialDirector: 16/0 ✅
- PopulationLedger: 24/0 ✅
- RevengeQuota: 18/0 ✅
- HordeManifest: 10/0 ✅

+++ mods/05-Rimconemy-Infected-Automation/Source/Story/StoryDirector.cs
@@ -1035,7 +1035,8 @@ namespace Rimconemy.InfectedAutomation.Story
             // rather than threading a negative through min/max.
             int freeBudgetRaw = ledger.Cap - ledger.HumanoidLiveCount;
             int freeBudget = (int)System.Math.Min(int.MaxValue, System.Math.Max(0, freeBudgetRaw));
-            int raw = (int)System.Math.Floor((double)ledger.RecentKillsToday * ratio);
+            // Float-Präzision: pure-float multiply bewahrt Ratio (0.7f → 7.0, nicht 6.999...).
+            int raw = (int)System.Math.Floor((float)ledger.RecentKillsToday * ratio);
             LastPendingRevenge = System.Math.Max(0, System.Math.Min(raw, freeBudget));
         }
